requirements.txt
txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
sentence-transformers==2.2.2
faiss-cpu==1.7.4
scikit-learn==1.4.2
pyyaml==6.0.2
joblib==1.4.2
flask==3.0.0
